Update test suite to work with Rspec 3 and newer ruby-saml
Since the last time that this gem was messed with the ruby-saml library
has changed from Onelogin::Saml to OneLogin::RubySaml. This updates the
uses of this gem to match the new namespace. It also updates the test
suite to work with rspec 3.0's "expect" syntax. In addition to this it
pins the gem versions so that the gems updating won't break the build
for anyone who might want to contribute to this project.