Add custom attributes_provider to encoded saml response.
In the event that a user would like to send custom attributes in the
assertion instead of the single email address attribute they can now
send in an attribute provider. This provider needs to implement a to_s
method that renders out the "AttributeStatement" element and associated
Attribute & AttributeValue elements.